[announce] Fixups
Pull the class out and the logic for going from text to  mode into a separate function to lower the cognitive complexity.

Use the display character as part of the underlying enum to avoid having them in separate places.

Use std::format and std::string instead of a static buffer.